(MAINT) Re-implement TOC
This change overhauls the implementation of the TOC and makes various
supporting changes.

It introduces the new `ShouldRenderToC` utility partial, which validates
that the settings for the page indicate it should render a TOC and that
the page has two or more headings. This is used as the preferred way to
check if a page should have the TOC rendered.

It completely reimplements the logic for parsing a page for headings to
support headings from render hooks, shortcodes, and HTML that are not
captured in the automatic TOC by Hugo. The new implementation discovers
the headings on a content page, parses them from HTML strings into maps
with their information, and then associates the child headings to their
parent.

This greatly simplifies the rendering logic for the TOC, allowing entries
to recurse instead of building a map of entries and closing them dynamically
during the render.

In the process of this reimplementation, the TOC now has two different
rendering options. The legacy option renders the TOC as an ordered or
unordered list of links, just like before. The new option renders the
TOC as the `<sl-tree>` element from Shoelace with each entry as an
`<sl-tree-item>`.

The new option includes several configuration options, including:

- Control over the indentation display for the TOC
- Setting alternate expand/collapse icons
- Defining prefix and suffix icons for an entry as heading attributes
- Defining whether the entry should be collapsed or expanded as a heading
  attribute

This change does not include documentation. Docs will be addressed in a future
commit for the entire changeset.