(MAINT) Reimplement mobile controls
This change reimplements the mobile controls for displaying the
site menu and page TOC, adding the `menu_control` and `toc_control`
settings to the `platen.display.mobile` site configuration option.

This update enables a user to select between the legacy template,
which displays the SVG icons already used in the site, and the new
template.

The new template replaces the SVG images with `<sl-icon-button>`
elements wrapped in a tooltip. With the new template, users can
configure the icon to display, the label for the button, and the
tooltip.

This change does not update the docs.