(MAINT) Update markup to support icon syntaxes
This change updates the implementation of the `alerts`, `buttons`, and
`details` markup to support the expanded options for icons. All three
now support the full set of options when defining icon options in YAML
and all of them now support the new shorthand syntax for icons.

This change updates the documentation for the markups to include examples
and clarify how they can take advantage of this improvement.