(MAINT) Add handling to fix TOC auto expand/collapse
This change adds the new `stopPropagationInTree` async function to the Platen
module and uses it in the new tocScript partial to ensure that selecting a
link in the TOC does not automatically expand or collapse that tree item.

This makes the behavior of the TOC more predictable and provides a built-in
way for other components to borrow this same behavior as needed.