(MAINT) Fix rendering TOC on pages with 1 heading
Prior to this change, Platen threw an arcane error when building the
TOC map for a page that only has a single heading. This change adds
special handling for single-heading pages to avoid the error and
render the TOC correctly.

This fixes an issue in the build for the `gh-dash` documentation.