(MAINT) Fix assets for non-root sites
Prior to this change, sites hosted with a Base URL that isn't set
to a root URL, like `https://example.com/shop` instead of just
`https://example.com`, raised errors for loading fonts and the
service worker.

This is because the URLs for the service worker and built-in
fontswere always pointing to `/` rather than the relative prefix.

This change:

1. Fixes the service worker templates to use the root of the Base URL
   instead of `/`
1. Adds a new convenience SCSS variable, `$BaseUrlPrefix`, to the
   SCSS. This variable can be used for interpolation.
1. Updates the built-in fonts style to use the `$BaseUrlPrefix`
   instead of looking for the fonts at `/...`.