(WIP) Tabs/Alpine rework for hexmap
This change builds on the prototype by @cyberselkie to
emit a hexmap on a page. It emits a list of buttons
above a shoelace tab group. The buttons should display
as hex map tiles. Clicking the buttons selects the
appropriate tab in the group. Eventually, selecting
a tab should also highlight/select the hex in the
map above.

None of the styling is in place. This PR is intended
for comparison only, the canonical PR should be the
original submission with updates and this PR and
commit should be thrown away at that time.